Use int for CheckValue.check()

+++ src/main/java/yokwe/majuro/type/CheckValue.java
@@ -1,5 +1,5 @@
 package yokwe.majuro.type;
 
 public interface CheckValue {
-	public void check(long value);
+	public void check(int value);
 }
